Start testing Cats' laws for Chimney types
