+++ chimney-cats/src/main/scala/io/scalaland/chimney/cats/package.scala
@@ -1,4 +1,7 @@
 package io.scalaland.chimney
 
 /** @since 0.5.0 */
-package object cats extends CatsTotalTransformerImplicits with CatsPartialTransformerImplicits
+package object cats
+    extends CatsTotalTransformerImplicits
+    with CatsPartialTransformerImplicits
+    with CatsPartialResultImplicits
