+++ chimney-cats/src/test/scala/io/scalaland/chimney/cats/PartialTransformerResultErrorLaws.scala
@@ -5,7 +5,7 @@ import io.scalaland.chimney.{partial, ChimneySpec}
 import org.scalacheck.Test.check
 import org.scalacheck.Prop.forAll
 
-class PartialTransformerResultErrorSemigroupSpec extends ChimneySpec {
+class PartialTransformerResultErrorLaws extends ChimneySpec {
 
   test("Semigroup[partial.Result.Errors] should aggregate errors from 2 partial.Result.Errors") {
     check(forAll { (str1: String, str2: String) =>
